[feat] commonizing component properties
