+++ frontend-web/src/api/widgetService.ts
@@ -6,7 +6,7 @@ const selectWidgetList = (): Promise<any> => get(URL_WIDGET);
 const selectWidget = (id: string, data = null): Promise<any> => get(URL_WIDGET + '/' + id, data);
 const createWidget = (data: unknown): Promise<any> => post(URL_WIDGET, data);
 const updateWidget = (data: unknown): Promise<any> => put(URL_WIDGET, data);
-const deleteWidget = (data: unknown): Promise<any> => del(URL_WIDGET, data);
+const deleteWidget = (id: string): Promise<any> => del(URL_WIDGET + '/' + id);
 
 const WidgetService = {
   selectWidgetList,
