+++ frontend-web/src/components/PageTitleBox.tsx
@@ -7,7 +7,6 @@ function PageTitleBox(props) {
   return (
     <Stack
       sx={{
-        flex: '1 1 auto',
         width: '100%',
       }}
     >
