+++ frontend-web/src/layouts/Layout.tsx
@@ -10,8 +10,8 @@ function Layout(props) {
   return (
     <Stack
       sx={{
-        flex: '1 1 auto',
-        // height: '100%',
+        width: '100%',
+        height: '100%',
       }}
     >
       <Header height={headerHeight} />
