+++ frontend-web/src/pages/Widget/index.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { Outlet, useParams } from 'react-router-dom';
+import { useAlert } from 'react-alert';
 import PageContainer from '@/components/PageContainer';
 import PageTitleBox from '@/components/PageTitleBox';
 import BoardList from '@/components/BoardList';
@@ -10,10 +11,10 @@ const title = '위젯';
 
 const Widget = () => {
   const { widgetId } = useParams();
+  const alert = useAlert();
 
   const [isLoading, setIsLoading] = useState(false);
   const [widgetList, setWidgetList] = useState([]);
-  // const [loadedCount, setLoadedCount] = useState(1);
 
   useEffect(() => {
     getWidgetList();
@@ -29,11 +30,30 @@ const Widget = () => {
     setIsLoading(true);
   };
 
+  const removeWidget = item => {
+    alert.success(`${item.title}\n위젯을 삭제하시겠습니까?`, {
+      title: '위젯 삭제',
+      closeCopy: '취소',
+      actions: [
+        {
+          copy: '삭제',
+          onClick: () => {
+            WidgetService.deleteWidget(item.id).then(response => {
+              if (response.status === 200) {
+                getWidgetList();
+              }
+            });
+          },
+        },
+      ],
+    });
+  };
+
   return (
     <PageContainer>
       {!widgetId ? (
         <PageTitleBox title={title} button={<AddIconButton link="create" />}>
-          <BoardList postList={widgetList} />
+          <BoardList postList={widgetList} handleDeleteSelect={removeWidget} />
         </PageTitleBox>
       ) : (
         <Outlet />
