+++ frontend-web/src/widget/modules/scatterchart/BubbleChart.tsx
@@ -1,22 +1,21 @@
 import React, { useEffect, useState } from 'react';
+import { Box } from '@mui/material';
 import ReactECharts from 'echarts-for-react';
-import { Box, Stack } from '@mui/material';
-import { getAggregationDataForChart, getGridSize, getLegendOption } from '@/modules/utils/chartUtil';
+import { getGridSize, getLegendOption } from '@/widget/modules/utils/chartUtil';
 
-const LineChart = props => {
-  const { option, dataSet, axis = 'x', seriesOp, defaultOp, createOp } = props;
-  const reverseAxis = axis === 'x' ? 'y' : 'x';
+function BubbleChart(props) {
+  const { option, dataSet, seriesOp, defaultOp, createOp } = props;
 
   const [componentOption, setComponentOption] = useState({});
 
   const defaultComponentOption = {
     grid: { top: 50, right: 50, bottom: 50, left: 50 },
     tooltip: { trigger: 'axis' },
-    [axis + 'Axis']: {
-      type: 'category',
+    xAxis: {
+      scale: true,
     },
-    [reverseAxis + 'Axis']: {
-      type: 'value',
+    yAxis: {
+      scale: true,
     },
     series: [],
     emphasis: {
@@ -29,7 +28,6 @@ const LineChart = props => {
   useEffect(() => {
     if (option && dataSet) {
       const newOption = createComponentOption();
-
       setComponentOption(newOption);
     }
   }, [option, dataSet]);
@@ -39,40 +37,29 @@ const LineChart = props => {
    * 위젯옵션과 데이터로
    * 컴포넌트에 맞는 형태로 생성
    */
+
   const createComponentOption = () => {
-    console.log('createComponentOption', option);
     let newOption = {};
 
     // series option에서 가져오기
     const newSeries = [];
-    let aggrData = [];
     option.series.forEach(item => {
-      console.log('dataSet', dataSet);
-      console.log('item', item);
-      console.log('axis', axis);
-      console.log('option[axisField]', option[axis + 'Field']);
-      aggrData = getAggregationDataForChart(dataSet, option[axis + 'Field'], item.field, item.aggregation);
-      console.log('aggrData : ', aggrData);
-      console.log(aggrData.map(dataItem => dataItem[item.field]));
-      if (item.field) {
+      if (item.xField || item.yField) {
         const series = {
-          name: item.field,
-          data: aggrData.map(dataItem => dataItem[item.field]),
-          type: item.type ? item.type : 'line',
+          type: 'scatter',
+          name: item.title,
+          data: dataSet.map(dataItem => [dataItem[item.xField], dataItem[item.yField], dataItem[item.symbolSize], 'text']),
+          symbolSize: function (data) {
+            return Math.sqrt(data[2]);
+          },
           color: item.color,
-          smooth: true,
           ...seriesOp,
         };
         newSeries.push(series);
       }
     });
-
-    if (aggrData && option[axis + 'Field']) {
+    if (dataSet) {
       const op = {
-        [axis + 'Axis']: {
-          type: 'category',
-          data: option[axis + 'Field'] ? aggrData.map(item => item[option[axis + 'Field']]) : '',
-        },
         series: newSeries,
         grid: getGridSize(option.legendPosition),
         legend: getLegendOption(option.legendPosition),
@@ -82,21 +69,19 @@ const LineChart = props => {
       newOption = { ...defaultComponentOption, ...op };
     }
 
-    // console.log(newOption);
     return newOption;
   };
 
   return (
-    <Stack
+    <Box
       sx={{
-        flex: '1 1 auto',
         width: '100%',
         height: '100%',
       }}
     >
       <ReactECharts option={componentOption} style={{ height: '100%', width: '100%' }} lazyUpdate={true} notMerge={true} />
-    </Stack>
+    </Box>
   );
-};
+}
 
-export default LineChart;
+export default BubbleChart;
