+++ frontend-web/src/widget/modules/utils/chartUtil.ts
