+++ frontend-web/src/widget/settings/LineChartSetting.tsx
@@ -42,7 +42,6 @@ const LineChartSetting = props => {
 
   return (
     <Grid item xs={10} md={4} lg={3} sx={{ display: 'flex', flexDirection: 'column' }}>
-      <WidgetTitleForm value={option.title} onChange={event => handleChange(event, setOption)} />
       <StyledList>
         <ListItem divider>
           <ListItemText primary="카테고리 설정" sx={{ textTransform: 'uppercase' }} />
